% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC}
\alias{QC}
\title{Allele Specific Expression Quality Control (ASEQC) Test}
\usage{
QC(ref, alt, numCores = 1)
}
\arguments{
\item{ref}{Dataframe containing reference count data}

\item{alt}{Dataframe containing alternate count data}

\item{numCores}{number of cores used for parallel execution (default = 1).}
}
\value{
Fitted std dev. and the ascertained quality cutoff.
}
\description{
This test is designed to detect poor quality ASE data. This includes
low coverage samples, samples with potential genotype mismatches while
generating the ASE data etc.
}
